Remove unwrapping handler
