+++ grpc/src/main/java/com/linecorp/armeria/client/grpc/GrpcClientBuilder.java
@@ -75,7 +75,6 @@ import com.linecorp.armeria.common.grpc.GrpcJsonMarshallerBuilder;
 import com.linecorp.armeria.common.grpc.GrpcSerializationFormats;
 import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageDeframer;
 import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageFramer;
-import com.linecorp.armeria.internal.common.grpc.UnwrappingGrpcExceptionHandlerFunction;
 import com.linecorp.armeria.unsafe.grpc.GrpcUnsafeBufferUtil;
 
 import io.grpc.CallCredentials;
@@ -419,8 +418,7 @@ public final class GrpcClientBuilder extends AbstractClientOptionsBuilder {
             option(INTERCEPTORS.newValue(clientInterceptors));
         }
         if (exceptionHandler != null) {
-            option(EXCEPTION_HANDLER.newValue(new UnwrappingGrpcExceptionHandlerFunction(
-                    exceptionHandler.orElse(GrpcExceptionHandlerFunction.of()))));
+            option(EXCEPTION_HANDLER.newValue(exceptionHandler.orElse(GrpcExceptionHandlerFunction.of())));
         }
 
         final Object client;
