+++ grpc/src/main/java/com/linecorp/armeria/client/grpc/GrpcClientOptions.java
@@ -37,7 +37,6 @@ import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageDeframer;
 import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageFramer;
 import com.linecorp.armeria.internal.client.grpc.NullCallCredentials;
 import com.linecorp.armeria.internal.client.grpc.NullGrpcClientStubFactory;
-import com.linecorp.armeria.internal.common.grpc.UnwrappingGrpcExceptionHandlerFunction;
 import com.linecorp.armeria.unsafe.grpc.GrpcUnsafeBufferUtil;
 
 import io.grpc.CallCredentials;
@@ -174,8 +173,7 @@ public final class GrpcClientOptions {
      * to a gRPC {@link Status}.
      */
     public static final ClientOption<GrpcExceptionHandlerFunction> EXCEPTION_HANDLER =
-            ClientOption.define("EXCEPTION_HANDLER", new UnwrappingGrpcExceptionHandlerFunction(
-                    GrpcExceptionHandlerFunction.of()));
+            ClientOption.define("EXCEPTION_HANDLER", GrpcExceptionHandlerFunction.of());
 
     /**
      * Sets whether to respect the marshaller specified in gRPC {@link MethodDescriptor}.
