+++ grpc/src/main/java/com/linecorp/armeria/internal/common/grpc/GrpcExceptionHandlerFunctionUtil.java
@@ -15,8 +15,12 @@
  */
 package com.linecorp.armeria.internal.common.grpc;
 
+import static java.util.Objects.requireNonNull;
+
 import com.linecorp.armeria.common.RequestContext;
 import com.linecorp.armeria.common.grpc.GrpcExceptionHandlerFunction;
+import com.linecorp.armeria.common.grpc.protocol.ArmeriaStatusException;
+import com.linecorp.armeria.common.util.Exceptions;
 
 import io.grpc.Metadata;
 import io.grpc.Status;
@@ -30,13 +34,34 @@ public final class GrpcExceptionHandlerFunctionUtil {
 
     public static Status fromThrowable(RequestContext ctx, GrpcExceptionHandlerFunction exceptionHandler,
                                        Throwable t, Metadata metadata) {
-        Status status = Status.fromThrowable(t);
+        final Status status = Status.fromThrowable(t);
         final Throwable cause = status.getCause();
-        if (cause != null) {
-            status = exceptionHandler.apply(ctx, status, cause, metadata);
+        if (cause == null) {
+            return status;
         }
+        return applyExceptionHandler(ctx, exceptionHandler, status, cause, metadata);
+    }
+
+    public static Status applyExceptionHandler(RequestContext ctx,
+                                               GrpcExceptionHandlerFunction exceptionHandler,
+                                               Status status, Throwable cause, Metadata metadata) {
+        final Throwable peeled = peelAndUnwrap(cause);
+        status = exceptionHandler.apply(ctx, status, peeled, metadata);
         assert status != null;
         return status;
+    }
+
+    private static Throwable peelAndUnwrap(Throwable t) {
+        requireNonNull(t, "t");
+        t = Exceptions.peel(t);
+        Throwable cause = t;
+        while (cause != null) {
+            if (cause instanceof ArmeriaStatusException) {
+                return StatusExceptionConverter.toGrpc((ArmeriaStatusException) cause);
+            }
+            cause = cause.getCause();
+        }
+        return t;
     }
 
     private GrpcExceptionHandlerFunctionUtil() {}
