+++ grpc/src/main/java/com/linecorp/armeria/server/grpc/HandlerRegistry.java
@@ -71,7 +71,6 @@ import com.linecorp.armeria.common.DependencyInjector;
 import com.linecorp.armeria.common.annotation.Nullable;
 import com.linecorp.armeria.common.grpc.GrpcExceptionHandlerFunction;
 import com.linecorp.armeria.internal.common.ReflectiveDependencyInjector;
-import com.linecorp.armeria.internal.common.grpc.UnwrappingGrpcExceptionHandlerFunction;
 import com.linecorp.armeria.internal.server.annotation.AnnotationUtil;
 import com.linecorp.armeria.internal.server.annotation.DecoratorAnnotationUtil;
 import com.linecorp.armeria.internal.server.annotation.DecoratorAnnotationUtil.DecoratorAndOrder;
@@ -282,8 +281,7 @@ final class HandlerRegistry {
             grpcExceptionHandler.ifPresent(exceptionHandler -> {
                 GrpcExceptionHandlerFunction grpcExceptionHandler0 = exceptionHandler;
                 if (defaultExceptionHandler != null) {
-                    grpcExceptionHandler0 = new UnwrappingGrpcExceptionHandlerFunction(
-                            exceptionHandler.orElse(defaultExceptionHandler));
+                    grpcExceptionHandler0 = exceptionHandler.orElse(defaultExceptionHandler);
                 }
                 grpcExceptionHandlersBuilder.put(methodDefinition, grpcExceptionHandler0);
             });
